break code into small components
