+++ src/App.jsx
@@ -1,11 +1,10 @@
 import { useState } from "react";
 import { nanoid } from "nanoid";
 import NewTodoForm from "./components/NewTodoForm";
-
+import TodoList from "./components/TodoList";
 import "./App.css";
 
 function App() {
-  // const [newItem, setNewItem] = useState("");
   const [todos, setTodos] = useState([]);
 
   function addTodo(title) {
@@ -13,21 +12,6 @@ function App() {
       return [...prevState, { id: nanoid(), title, completed: false }];
     });
   }
-
-  // Keeps track of changes in the form input
-  // function handleChange(e) {
-  //   setNewItem(e.target.value);
-  // }
-
-  // function handleSubmit(e) {
-  //   e.preventDefault();
-
-  //   setTodos((prevState) => {
-  //     return [...prevState, { id: nanoid(), title: newItem, completed: false }];
-  //   });
-
-  //   setNewItem("");
-  // }
 
   // toggles checked props for todo items
   function toggleTodo(id, completed) {
@@ -41,48 +25,50 @@ function App() {
     });
   }
 
-  // Generate JSX and display todo items
-  const todoListItems = todos.map((todo) => {
-    // if item is checked, font turns green
-    const styles = {
-      color: todo.completed ? "green" : "white",
-    };
+  // Delete Item
+  function deleteTodo(id) {
+    setTodos((prevState) => {
+      return prevState.filter((todo) => todo.id !== id);
+    });
+  }
 
-    // Delete Item
-    function deleteTodo(id) {
-      setTodos((prevState) => {
-        return prevState.filter((todo) => todo.id !== id);
-      });
-    }
+  // ! Generate JSX and display todo items DO NOT USE
+  // ! Code has been refactored and broken up into smaller components
+  // ! Check TodoList.jsx and TodoItem.jsx
+  // const todoListItems = todos.map((todo) => {
+  //   // if item is checked, font turns green
+  //   const styles = {
+  //     color: todo.completed ? "green" : "white",
+  //   };
 
-    return (
-      <li key={todo.id}>
-        <label style={styles}>
-          <input
-            type="checkbox"
-            checked={todo.completed}
-            onChange={(e) => toggleTodo(todo.id, e.target.checked)}
-          />
-          {todo.title}
-        </label>
-        <button onClick={() => deleteTodo(todo.id)} className="btn btn-delete">
-          Delete
-        </button>
-      </li>
-    );
-  });
+  //   return (
+  //     <li key={todo.id}>
+  //       <label style={styles}>
+  //         <input
+  //           type="checkbox"
+  //           checked={todo.completed}
+  //           onChange={(e) => toggleTodo(todo.id, e.target.checked)}
+  //         />
+  //         {todo.title}
+  //       </label>
+  //       <button onClick={() => deleteTodo(todo.id)} className="btn btn-delete">
+  //         Delete
+  //       </button>
+  //     </li>
+  //   );
+  // });
 
   return (
     <main>
       <h1>TODO LIST APP</h1>
-      {/* New Item Form */}
+
       <NewTodoForm addTodo={addTodo} />
+
       <h1 className="header">Todo List</h1>
+
       {/* Todo List Items */}
-      <ul className="list">
-        {todos.length === 0 && "No Todos"}
-        {todoListItems}
-      </ul>
+
+      <TodoList todos={todos} toggleTodo={toggleTodo} deleteTodo={deleteTodo} />
     </main>
   );
 }
