+++ src/components/TodoItem.jsx
@@ -1,0 +1,28 @@
+import React from "react";
+
+export default function TodoItem({
+  completed,
+  id,
+  title,
+  toggleTodo,
+  deleteTodo,
+}) {
+  const styles = {
+    color: completed ? "green" : "white",
+  };
+  return (
+    <li key={id}>
+      <label style={styles}>
+        <input
+          type="checkbox"
+          checked={completed}
+          onChange={(e) => toggleTodo(id, e.target.checked)}
+        />
+        {title}
+      </label>
+      <button onClick={() => deleteTodo(id)} className="btn btn-delete">
+        Delete
+      </button>
+    </li>
+  );
+}
