+++ src/components/TodoList.jsx
@@ -1,5 +1,22 @@
 import React from "react";
+import TodoItem from "./TodoItem";
 
-export default function TodoList(props) {
-  return <h1>Todo List</h1>;
+export default function TodoList({ todos, toggleTodo, deleteTodo }) {
+  console.log(todos);
+  return (
+    <ul className="list">
+      {todos.length === 0 && "No Todos"}
+
+      {todos.map((todo) => {
+        return (
+          <TodoItem
+            {...todo}
+            key={todo.id}
+            toggleTodo={toggleTodo}
+            deleteTodo={deleteTodo}
+          />
+        );
+      })}
+    </ul>
+  );
 }
